move mb_fig to plotting.py & fix label position
